pyproject.toml
[project]
name = "reference-openadr-bl-gac"
version = "0.1.0"
description = ""
authors = [
    {name = "Nick van der Burgt",email = "nick.van.der.burgt@elaad.nl"}
]
readme = "README.md"
requires-python = ">=3.12, <4"
dependencies = [
    "openadr3-client (>=0.0.6,<0.0.7)",
    "openadr3-client-gac-compliance (>=1.2.1,<2.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "influxdb-client[async] (>=1.49.0,<2.0.0)",
    "ruff (>=0.12.4,<0.13.0)",
    "mypy (>=1.17.0,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
# We do not have to package our API, so disable packaging.
package-mode = false